+++ src/index.ts
@@ -1,5 +1,5 @@
 export { NzConnection, NzConnectionConfig } from './NzConnection';
-export type { QueryResult, ExecuteResult } from './NzConnection';
+export type { QueryResult, QueryResultRow, ExecuteResult } from './NzConnection';
 export { NzCommand } from './NzCommand';
 export { NzDataReader, type ColumnDescription, type ColumnMetadata, type GeneratorItem } from './NzDataReader';
 export { NzPool } from './NzPool';